+++ src/server/api/routers/WishlistController.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { createTRPCRouter, publicProcedure } from "../trpc";
-import { WishlistService } from "../../../services/WishlistService";
+import { WishlistService } from "../../services/WishlistService";
 
 const wishlistService = new WishlistService();
