+++ src/server/services/Event.ts
@@ -1,6 +1,6 @@
 import { db as prisma } from "~/server/db";
-import type { Event, User } from "@prisma/client";
-import { Status } from "@prisma/client"
+import type { Event } from "@prisma/client";
+import { StatusType } from "@prisma/client";
 
 export class EventEntity {
   private data: Event;
@@ -8,19 +8,19 @@ export class EventEntity {
   constructor(event: Event) {
     this.data = event;
   }
-
+/*  Depricated
   static async publishEvent(eventId: number): Promise<void> {
     const event = await prisma.event.findUnique({ where: { id: eventId } });
     if (!event) throw new Error("Event not found");
     // Logic to make event visible, if needed
   }
-
+*/
   async addGuest(guestId: string): Promise<void> {
     await prisma.invitation.create({
       data: {
         eventId: this.data.id,
-        guestId: guestId,
-        status: Status.PENDING,
+        guestUsername: guestId,
+        status: StatusType.PENDING,
         
       },
     });
