+++ src/server/services/EventPlanner.ts
@@ -1,8 +1,10 @@
 import { db as prisma } from "~/server/db";
 import { EventEntity } from "./Event";
-import { Status } from "@prisma/client";
+import { StatusType } from "@prisma/client";
 import { generateToken } from "~/utils/token";
 import { EventManagementException } from "./EventManagementException";
+import { nanoid } from "nanoid"
+
 
 export class EventPlanner {
   async createEvent(data: {
@@ -20,8 +22,15 @@ export class EventPlanner {
         location: data.location,
         date: data.date,
         time: data.time,
-        token: generateToken(),
-        createdByUsername: data.createdBy,
+	token: nanoid(12),
+        //createdByUsername: data.createdBy,
+	user: {
+        //connect: { username: data.createdBy },
+        connectOrCreate: {
+    where: { username: data.createdBy },
+    create: { username: data.createdBy },
+  },
+	},
       },
     });
 
@@ -32,26 +41,53 @@ export class EventPlanner {
     await prisma.event.delete({ where: { id: eventId } });
   }
 
-  async sendInvitation(eventId: number, guestUsername: string): Promise<void> {
-    const exists = await prisma.user.findUnique({ 
-      where: { username: guestUsername } 
-    });
-    
+  async sendInvitation(eventId: number, guestId: string): Promise<void> {
+    const exists = await prisma.user.findUnique({ where: { username: guestId } });
     if (!exists) throw new EventManagementException("Guest does not exist");
 
     await prisma.invitation.create({
       data: {
         eventId: eventId,
-        guestUsername: guestUsername,
-        status: Status.PENDING,
+        guestUsername: guestId,
+        status: StatusType.PENDING,
+        createdAt: new Date(),
+	//guest:{ connect: { username: guestId }},
+	//event:{ connect: { id: eventId } },
       },
     });
   }
-
-  async getEventGuests(eventId: number) {
-    return await prisma.invitation.findMany({
+/* Depricated
+  async manageWishlist(eventId: number) {
+    const wishlist = await prisma.eventItem.findMany({
       where: { eventId: eventId },
-      include: { guest: true },
+      include: { item: true },
     });
+    return wishlist;
   }
+
+  async viewAnalytics(eventId: number) {
+    const inviteCount = await prisma.invitation.count({ where: { eventId: eventId } });
+    const accepted = await prisma.invitation.count({
+      where: { eventId: eventId, status: Status.ACCEPTED },
+    });
+    const declined = await prisma.invitation.count({
+      where: { eventId: eventId, status: Status.REJECTED },
+    });
+
+    return {
+      inviteCount,
+      accepted,
+      declined,
+    };
+  }
+
+  async manageGallery(eventId: number) {
+    return await prisma.media.findMany({ where: { eventId: eventId } });
+  }
+
+  async receiveContribution(eventId: number) {
+    return await prisma.contribution.findMany({ where: { eventId: eventId } });
+  }
+*/
+
 }
