--- src/services/EventManagementException.ts
@@ -1,6 +1,0 @@
-export class EventManagementException extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "EventManagementException";
-  }
-}
